pyproject.toml
[project]
name = "lookup"
version = "0.1.0"
description = ""
authors = [
    {name = "BenjaminAlejandroZepeda",email = "ben.zepeda@duocuc.cl"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "supabase (>=2.28.2,<3.0.0)",
    "pandas (>=3.0.2,<4.0.0)",
    "kaggle (>=2.0.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.poetry]
package-mode = false